Restructure dep_manager and separate container, redis, and neo4j from central code. Load containers in foreground, will change to background later
