+++ beeflow/common/deps/neo4j_manager.py
@@ -1,0 +1,187 @@
+import socket
+import os
+import shutil
+import re
+import subprocess
+import time
+
+from beeflow.wf_manager.resources import wf_utils
+from beeflow.common.db import wfm_db
+
+from beeflow.common import paths
+from beeflow.common.deps import container_manager
+
+from beeflow.common.config_driver import BeeConfig as bc
+from beeflow.common import log as bee_logging
+
+# Define directories within module scope
+bee_workdir = paths._workdir()
+mount_dir = os.path.join(bee_workdir, 'gdb_mount')
+data_dir = mount_dir + '/data'
+logs_dir = mount_dir + '/logs'
+run_dir = mount_dir + '/run'
+certs_dir = mount_dir + '/certificates'
+confs_dir = mount_dir + "/conf"
+container_path = container_manager.get_container_dir('neo4j')
+log = bee_logging.setup('neo4j')
+
+
+def get_open_port():
+    """Return an open ephemeral port."""
+    s = socket.socket(socket.AF_INET, socket.SOCK_STREAM)
+    s.bind(("", 0))
+    s.listen(1)
+    port = s.getsockname()[1]
+    s.close()
+    return port
+
+
+def setup_ports():
+    """Return three open ports for bolt, http, https"""
+    # Get ports for neo4j to run
+    bolt_port = get_open_port()
+    http_port = get_open_port()
+    https_port = get_open_port()
+
+    db = wf_utils.connect_db(wfm_db, wf_utils.get_db_path())
+
+    db.info.set_port('bolt', bolt_port)
+    db.info.set_port('http', http_port)
+    db.info.set_port('https', https_port)
+    return bolt_port, http_port, https_port
+
+
+def setup_mounts():
+    """Set up mount directories for the graph database."""
+
+    os.makedirs(data_dir, exist_ok=True)
+    os.makedirs(logs_dir, exist_ok=True)
+    os.makedirs(mount_dir, exist_ok=True)
+    os.makedirs(certs_dir, exist_ok=True)
+    os.makedirs(run_dir, exist_ok=True)
+
+
+def setup_configs(bolt_port, http_port, https_port):
+    """Set up GDB configuration.
+
+    This function needs to be run before each new invocation since the
+    config file could have changed.
+    """
+
+    os.makedirs(confs_dir, exist_ok=True)
+    gdb_configfile = shutil.copyfile(container_path + "/var/lib/neo4j/conf/neo4j.conf",
+                                     confs_dir + "/neo4j.conf")
+    log.info(gdb_configfile)
+
+    with open(gdb_configfile, "rt", encoding="utf8") as cfile:
+        data = cfile.read()
+
+    log.info(f"new bolt: {bolt_port}")
+    bolt_config = r'#(server.bolt.listen_address=):[0-9]*'
+    data = re.sub(bolt_config, rf'\1:{bolt_port}', data)
+    http_config = r'#(server.http.listen_address=):[0-9]*'
+    data = re.sub(http_config, rf'\1:{http_port}', data)
+    https_config = r'#(server.https.listen_address=):[0-9]*'
+
+    with open(gdb_configfile, "wt", encoding="utf8") as cfile:
+        cfile.write(data)
+def create_credentials():
+    """Create the password and set the logfiles in environment"""
+    db_password = bc.get('graphdb', 'dbpass')
+    try:
+        command = ['neo4j-admin', 'dbms', 'set-initial-password', str(db_password)]
+        subprocess.run([
+            "ch-run",
+            "--set-env=" + container_path + "/ch/environment",
+            "-b", confs_dir + ":/var/lib/neo4j/conf",
+            "-b", data_dir + ":/data",
+            "-b", logs_dir + ":/logs",
+            "-b", run_dir + ":/var/lib/neo4j/run", container_path,
+            "--", *command
+        ], check=True)
+    except subprocess.CalledProcessError:
+        log.error("neo4j-admin set-initial-password failed")
+
+def create_database():
+    """Create the neo4j database and return the process"""
+    try:
+        command = ['neo4j', 'console']
+        proc = subprocess.Popen(["ch-run",
+                              "--set-env=" + container_path + "/ch/environment",
+                               "-b",
+                               confs_dir + ":/var/lib/neo4j/conf",
+                               "-b",
+                               data_dir + ":/data",
+                               "-b",
+                               logs_dir + ":/logs",
+                               "-b",
+                               run_dir + ":/var/lib/neo4j/run",
+                               "-b",
+                               certs_dir + ":/var/lib/neo4j/certificates",
+                               container_path,
+                               "--", *command
+                               ], stdout=subprocess.PIPE, stderr=subprocess.PIPE)
+        wait_gdb()
+        return proc
+    except FileNotFoundError:
+        log.error("Neo4j failed to start.")
+        return -1
+
+
+def start():
+    """Start the graph database."""
+    log.info('Starting Neo4j Database')
+
+    bolt_port, http_port, https_port = setup_ports()
+
+    setup_configs(bolt_port, http_port, https_port)
+
+    setup_mounts()
+
+    create_credentials()
+
+    return create_database()
+
+def start_gdb():
+    """
+    Function to start the neo4j graph database
+    Creates the database image and starts the database
+    """
+
+
+    try:
+        create_image()
+    except NoContainerRuntime:
+        crt_message = "Charliecloud not installed in current environment."
+        dep_log.info(crt_message)
+        return -1
+    
+        proc = start_neo4j(mount_dir, bolt_port, http_port,
+            https_port)
+
+    wait_gdb()
+    return proc
+
+
+
+def wait_gdb():
+    """Need to wait for the GDB. Currently, we're using the sleep time paramater.
+
+    We'd like to remove that in the future.
+    """
+    gdb_sleep_time = bc.get('graphdb', 'sleep_time')
+    print(f'waiting {gdb_sleep_time}s for GDB to come up')
+    time.sleep(gdb_sleep_time)
+
+
+def remove_gdb():
+    """Remove the current GDB bind mount directory."""
+    bee_workdir = wf_utils.get_bee_workdir()
+    gdb_workdir = os.path.join(bee_workdir, 'current_gdb')
+    old_gdb_workdir = os.path.join(bee_workdir, 'old_gdb')
+    if os.path.isdir(gdb_workdir):
+        # Rename the directory to guard against NFS errors
+        shutil.move(gdb_workdir, old_gdb_workdir)
+        time.sleep(2)
+        shutil.rmtree(old_gdb_workdir)
+        time.sleep(2)
