--- beeflow/wf_manager/common/dep_manager.py
@@ -1,272 +1,0 @@
-#!/usr/bin/env python3
-
-"""Functions for managing the BEE depency container and associated bind mounts."""
-
-import os
-import re
-import time
-import shutil
-import signal
-import socket
-import subprocess
-
-from beeflow.common import log as bee_logging
-from beeflow.common.db.bdb import connect_db
-from beeflow.common.db import wfm_db
-from beeflow.wf_manager.resources import wf_utils
-from beeflow.common.config_driver import BeeConfig as bc
-
-from celery import shared_task #noqa pylama can't find celery
-
-
-dep_log = bee_logging.setup(__name__)
-
-
-class NoContainerRuntime(Exception):
-    """An exception for no container runtime like charliecloud or singularity."""
-
-
-def check_container_runtime():
-    """Check if the container runtime is currently installed."""
-    # Needs to support singuarity as well
-    if shutil.which("ch-convert") is None or shutil.which("ch-run") is None:
-        dep_log.error("ch-convert or ch-run not found. Charliecloud required"
-                      " for neo4j container.")
-        raise NoContainerRuntime('')
-
-
-def make_dep_dir():
-    """Make a new bee dependency container directory."""
-    bee_workdir = wf_utils.get_bee_workdir()
-    bee_dir = f'{bee_workdir}/deps'
-    bee_dir_exists = os.path.isdir(bee_dir)
-    if not bee_dir_exists:
-        os.makedirs(bee_dir)
-
-
-def get_dep_dir():
-    """Return the dependency directory path."""
-    bee_workdir = wf_utils.get_bee_workdir()
-    bee_container_dir = f'{bee_workdir}/deps/'
-    return bee_container_dir
-
-
-def get_container_dir():
-    """Return the depency container path."""
-    container_name = 'dep_container'
-    return get_dep_dir() + container_name
-
-
-def check_container_dir():
-    """Return true if the container directory exists."""
-    container_dir = get_container_dir()
-    container_dir_exists = os.path.isdir(container_dir)
-    return container_dir_exists
-
-def get_open_port():
-    """Return an open ephemeral port."""
-    s = socket.socket(socket.AF_INET, socket.SOCK_STREAM)
-    s.bind(("", 0))
-    s.listen(1)
-    port = s.getsockname()[1]
-    s.close()
-    return port
-
-
-def setup_gdb_mounts(mount_dir):
-    """Set up mount directories for the graph database."""
-    data_dir = mount_dir + '/data'
-    logs_dir = mount_dir + '/logs'
-    run_dir = mount_dir + '/run'
-    certs_dir = mount_dir + '/certificates'
-
-    data_dir = os.path.join(data_dir, "data")
-    os.makedirs(data_dir, exist_ok=True)
-
-    logs_dir = os.path.join(logs_dir, "logs")
-    os.makedirs(logs_dir, exist_ok=True)
-
-    mount_dir = os.path.join(run_dir, "run")
-    os.makedirs(mount_dir, exist_ok=True)
-
-    gdb_certs_path = os.path.join(certs_dir, "certificates")
-    os.makedirs(gdb_certs_path, exist_ok=True)
-
-
-def setup_gdb_configs(mount_dir, bolt_port, http_port, https_port):
-    """Set up GDB configuration.
-
-    This function needs to be run before each new invocation since the
-    config file could have changed.
-    """
-    container_path = get_container_dir()
-    confs_dir = os.path.join(mount_dir, "conf")
-    os.makedirs(confs_dir, exist_ok=True)
-    gdb_configfile = shutil.copyfile(container_path + "/var/lib/neo4j/conf/neo4j.conf",
-                                     confs_dir + "/neo4j.conf")
-    dep_log.debug(gdb_configfile)
-
-    with open(gdb_configfile, "rt", encoding="utf8") as cfile:
-        data = cfile.read()
-
-    bolt_config = r'#(server.bolt.listen_address=):[0-9]*'
-    data = re.sub(bolt_config, rf'\1:{bolt_port}', data)
-    http_config = r'#(server.http.listen_address=):[0-9]*'
-    data = re.sub(http_config, rf'\1:{http_port}', data)
-    https_config = r'#(server.https.listen_address=):[0-9]*'
-    data = re.sub(https_config, rf'\1:{https_port}', data)
-    with open(gdb_configfile, "wt", encoding="utf8") as cfile:
-        cfile.write(data)
-
-
-def create_image():
-    """Create a new BEE dependency container if one does not exist.
-
-    By default, the container is stored in /tmp/<user>/beeflow/deps.
-    """
-    # Can throw an exception that needs to be handled by the caller
-    check_container_runtime()
-
-    dep_img = bc.get('DEFAULT', 'neo4j_image')
-
-    # Check for BEE dependency container directory:
-    container_dir_exists = check_container_dir()
-    if container_dir_exists:
-        dep_log.info('Already have neo4j container')
-        return
-
-    make_dep_dir()
-    container_dir = get_container_dir()
-    # Build new dependency container
-    try:
-        subprocess.run(["ch-convert", "-i", "tar", "-o", "dir",
-                        str(dep_img), str(container_dir)], check=True)
-    except subprocess.CalledProcessError as error:
-        dep_log.error(f"ch-convert failed: {error}")
-        shutil.rmtree(container_dir)
-        dep_log.debug(f"GraphDB container mount directory {container_dir} removed")
-        return
-
-    # Make the certificates directory
-    container_certs_path = os.path.join(container_dir, 'var/lib/neo4j/certificates')
-    os.makedirs(container_certs_path, exist_ok=True)
-
-
-def start_neo4j(mount_dir, bolt_port, http_port, https_port, reexecute=False):
-    """Start the graph database."""
-    # We need to rerun the mount step before each start
-
-    setup_gdb_configs(mount_dir, bolt_port, http_port, https_port)
-    if not reexecute:
-        setup_gdb_mounts(mount_dir)
-
-    db_password = bc.get('graphdb', 'dbpass')
-    data_dir = mount_dir + '/data'
-    logs_dir = mount_dir + '/logs'
-    run_dir = mount_dir + '/run'
-    certs_dir = mount_dir + '/certificates'
-    confs_dir = mount_dir + "/conf"
-
-    container_path = get_container_dir()
-    if not reexecute:
-        try:
-            command = ['neo4j-admin', 'dbms', 'set-initial-password', str(db_password)]
-            subprocess.run([
-                "ch-run",
-                "--set-env=" + container_path + "/ch/environment",
-                "-b", confs_dir + ":/var/lib/neo4j/conf",
-                "-b", data_dir + ":/data",
-                "-b", logs_dir + ":/logs",
-                "-b", run_dir + ":/var/lib/neo4j/run", container_path,
-                "--", *command
-            ], check=True)
-        except subprocess.CalledProcessError:
-            dep_log.error("neo4j-admin set-initial-password failed")
-            return -1
-
-    try:
-        command = ['neo4j', 'console']
-        proc = subprocess.Popen(["ch-run",
-                              "--set-env=" + container_path + "/ch/environment",
-                               "-b",
-                               confs_dir + ":/var/lib/neo4j/conf",
-                               "-b",
-                               data_dir + ":/data",
-                               "-b",
-                               logs_dir + ":/logs",
-                               "-b",
-                               run_dir + ":/var/lib/neo4j/run",
-                               "-b",
-                               certs_dir + ":/var/lib/neo4j/certificates",
-                               container_path,
-                               "--", *command
-                               ], stdout=subprocess.PIPE, stderr=subprocess.PIPE)
-        return proc
-    except FileNotFoundError:
-        dep_log.error("Neo4j failed to start.")
-        return -1
-
-def start_gdb():
-    """
-    Function to start the neo4j graph database
-    Creates the database image and starts the database
-    """
-    dep_log.info("Starting GDB")
-    bolt_port = get_open_port()
-    http_port = get_open_port()
-    https_port = get_open_port()
-    
-    db = wf_utils.connect_db(wfm_db, wf_utils.get_db_path())
-
-    db.info.set_port('bolt', bolt_port)
-    db.info.set_port('http', http_port)
-    db.info.set_port('https', https_port)
-
-    try:
-        create_image()
-    except NoContainerRuntime:
-        crt_message = "Charliecloud not installed in current environment."
-        dep_log.info(crt_message)
-        return -1
-    
-    bee_workdir = wf_utils.get_bee_workdir()
-    mount_dir = os.path.join(bee_workdir, 'gdb_mount')
-    proc = start_neo4j(mount_dir, bolt_port, http_port,
-            https_port)
-
-    wait_gdb()
-    return proc
-
-
-
-def wait_gdb():
-    """Need to wait for the GDB. Currently, we're using the sleep time paramater.
-
-    We'd like to remove that in the future.
-    """
-    gdb_sleep_time = bc.get('graphdb', 'sleep_time')
-    print(f'waiting {gdb_sleep_time}s for GDB to come up')
-    time.sleep(gdb_sleep_time)
-
-
-def remove_gdb():
-    """Remove the current GDB bind mount directory."""
-    bee_workdir = wf_utils.get_bee_workdir()
-    gdb_workdir = os.path.join(bee_workdir, 'current_gdb')
-    old_gdb_workdir = os.path.join(bee_workdir, 'old_gdb')
-    if os.path.isdir(gdb_workdir):
-        # Rename the directory to guard against NFS errors
-        shutil.move(gdb_workdir, old_gdb_workdir)
-        time.sleep(2)
-        shutil.rmtree(old_gdb_workdir)
-        time.sleep(2)
-
-
-def kill_gdb(pid):
-    """Kill the GDB with the associated pid."""
-    try:
-        os.kill(pid, signal.SIGTERM)
-    except OSError:
-        dep_log.info('Process already killed')
-        return False
-    return True
