+++ beeflow/wf_manager/resources/wf_update.py
@@ -10,7 +10,6 @@ import jsonpickle
 from flask import make_response, jsonify
 from flask_restful import Resource, reqparse
 from beeflow.wf_manager.resources import wf_utils
-from beeflow.wf_manager.common import dep_manager
 from beeflow.common import log as bee_logging
 
 from beeflow.common.db import wfm_db
